working ebean login and registration
