--- src/main/java/model/entity/Entity.java
@@ -1,9 +1,0 @@
-package model.entity;
-
-/**
- * Created by adam on 26/02/2017.
- */
-public interface Entity {
-    public Integer getId();
-    public void setId(Integer id);
-}
