+++ src/main/java/org/cba/controller/DefaultController.java
@@ -1,4 +1,4 @@
-package controller;
+package org.cba.controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
