+++ src/main/java/org/cba/controller/ErrorController.java
@@ -1,4 +1,4 @@
-package controller;
+package org.cba.controller;
 
 import hyggemvc.controller.Controller;
