+++ src/main/java/org/cba/model/exception/EmailTakenException.java
@@ -1,4 +1,4 @@
-package model.exception;
+package org.cba.model.exception;
 
 /**
  * Created by adam on 28/02/2017.
