+++ src/main/resources/ebean.properties
@@ -1,0 +1,21 @@
+
+## load external properties
+#load.properties.override=${user.home}/config/my-application.properties
+
+# the name of the default server
+datasource.default=db
+
+
+## define these in external properties ...
+
+datasource.db.username=root
+datasource.db.password=root
+datasource.db.databaseUrl=jdbc:mysql://127.0.0.1:3306/cba_carport
+datasource.db.databaseDriver=com.mysql.jdbc.Driver
+
+#ebean.docstore.url=http://127.0.0.1:9200
+#ebean.docstore.active=true
+##ebean.docstore.generateMapping=true
+##ebean.docstore.dropCreate=true
+##ebean.docstore.create=true
+
