+++ src/test/resources/test-ebean.properties
@@ -1,0 +1,28 @@
+
+## when running tests use H2 and create a new DB
+
+ebean.ddl.generate=true
+ebean.ddl.run=true
+ebean.ddl.createOnly=false
+
+datasource.default=db
+
+datasource.db.username=root
+datasource.db.password=root
+datasource.db.databaseUrl=jdbc:mysql://127.0.0.1:3306/cba_carport_test
+datasource.db.databaseDriver=com.mysql.jdbc.Driver
+
+#datasource.db.username=
+#datasource.db.password=
+#datasource.db.databaseUrl=jdbc:sqlite:mydb.db
+#datasource.db.databaseDriver=org.sqlite.JDBC
+#datasource.db.isolationlevel=read_uncommitted
+
+
+#ebean.docstore.url=http://127.0.0.1:9200
+#ebean.docstore.active=true
+#ebean.docstore.generateMapping=true
+#ebean.docstore.dropCreate=true
+#ebean.docstore.create=true
+
+
