+++ web/WEB-INF/web.xml
@@ -5,7 +5,7 @@
          version="3.1">
     <filter>
         <filter-name>RequestDispatcher</filter-name>
-        <filter-class>servlet.RequestDispatcher</filter-class>
+        <filter-class>org.cba.servlet.RequestDispatcher</filter-class>
     </filter>
     <filter-mapping>
         <filter-name>RequestDispatcher</filter-name>
